Move import vault to core and reuse it in both clients
