+++ core/ui/navigation/sharedViews.tsx
@@ -1,5 +1,6 @@
 import { AddressPage } from '@core/ui/chain/components/address/AddressPage'
 import { ReshareVaultPage } from '@core/ui/mpc/keygen/reshare/ReshareVaultPage'
+import { CoreViewId } from '@core/ui/navigation/CoreView'
 import { CurrencyPage } from '@core/ui/preferences/currency'
 import { DefaultChainsPage } from '@core/ui/preferences/default-chains'
 import { LanguagePage } from '@core/ui/preferences/language'
@@ -8,6 +9,7 @@ import { VaultChainCoinPage } from '@core/ui/vault/chain/coin/VaultChainCoinPage
 import { ManageVaultChainsPage } from '@core/ui/vault/chain/manage'
 import { ManageVaultChainCoinsPage } from '@core/ui/vault/chain/manage/coin/ManageVaultChainCoinsPage'
 import { VaultChainPage } from '@core/ui/vault/chain/VaultChainPage'
+import { ImportVaultPage } from '@core/ui/vault/import/components/ImportVaultPage'
 import { NewVaultPage } from '@core/ui/vault/new'
 import { SendPage } from '@core/ui/vault/send/SendPage'
 import { VaultDetailsPage } from '@core/ui/vault/settings/details'
@@ -15,12 +17,11 @@ import { VaultRenamePage } from '@core/ui/vault/settings/rename'
 import { SwapPage } from '@core/ui/vault/swap/components/SwapPage'
 import { Views } from '@lib/ui/navigation/Views'
 
-import { CoreViewId } from './CoreView'
-
 type SharedViewId = Extract<
   CoreViewId,
   | 'currencySettings'
   | 'defaultChains'
+  | 'importVault'
   | 'languageSettings'
   | 'newVault'
   | 'renameVault'
@@ -36,12 +37,19 @@ type SharedViewId = Extract<
 >
 
 export const sharedViews: Views<SharedViewId> = {
+  address: AddressPage,
   currencySettings: CurrencyPage,
   defaultChains: DefaultChainsPage,
+  importVault: ImportVaultPage,
   languageSettings: LanguagePage,
   manageVaultChains: () => (
     <ActiveVaultGuard>
       <ManageVaultChainsPage />
+    </ActiveVaultGuard>
+  ),
+  manageVaultChainCoins: () => (
+    <ActiveVaultGuard>
+      <ManageVaultChainCoinsPage />
     </ActiveVaultGuard>
   ),
   newVault: NewVaultPage,
@@ -53,11 +61,6 @@ export const sharedViews: Views<SharedViewId> = {
   reshareVault: () => (
     <ActiveVaultGuard>
       <ReshareVaultPage />
-    </ActiveVaultGuard>
-  ),
-  vaultDetails: () => (
-    <ActiveVaultGuard>
-      <VaultDetailsPage />
     </ActiveVaultGuard>
   ),
   send: () => (
@@ -80,10 +83,9 @@ export const sharedViews: Views<SharedViewId> = {
       <VaultChainCoinPage />
     </ActiveVaultGuard>
   ),
-  manageVaultChainCoins: () => (
+  vaultDetails: () => (
     <ActiveVaultGuard>
-      <ManageVaultChainCoinsPage />
+      <VaultDetailsPage />
     </ActiveVaultGuard>
   ),
-  address: () => <AddressPage />,
 }
